requirements.txt
asgiref==3.5.0
attrs==21.4.0
autopep8==1.6.0
coverage==6.3.1
Django==4.0.2
Faker==12.0.1
flake8==4.0.1
iniconfig==1.1.1
mccabe==0.6.1
mixer==7.2.1
packaging==21.3
Pillow==9.0.1
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.7
pytest==7.0.1
pytest-cov==3.0.0
pytest-django==4.5.2
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
